core: Add thread related definitions
With pthread-embedded library, special care has to be taken when dealing with
conversions from `pthread_t` to scalar values. Therefore we have to define our
own conversion functions.

We also copy `ruby_fill_thread_id_string()` function from origin to glue in
order to avoid patching the origin code. The function is copied here because of
a bug in the Ruby implementation, which only defines it if THREAD_DEBUG is
enabled. When this bug is fixed in Ruby, this is no long necessary.

Signed-off-by: Costin Lupu <costin.lupu@cs.pub.ro>
Reviewed-by: Stefan Teodorescu <stefanl.teodorescu@gmail.com>