extensions: Add generated files
Each extensions has its own <extconf.h> header, which is generated based on the
build system configuration. We choose to add this header for `socket` extension
into the glue code because we need to change some of its configurations in the
next patch. All the extensions are registered in the `extinit.c` file.

Signed-off-by: Costin Lupu <costin.lupu@cs.pub.ro>
Reviewed-by: Stefan Teodorescu <stefanl.teodorescu@gmail.com>